# app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Couchbase Query Analyzer API
  version: 4.0.0-Beta
  description: |
    REST API for the **Couchbase Query Analyzer Server Edition**.

    All persistence is handled by the embedded **Couchbase Lite (CE)** store
    (`cbl_store.py`) — there is no external Couchbase Server dependency.
    Source data (`system:completed_requests` JSON) is provided by the user
    via paste / file upload in the UI; there is no live N1QL fetch.

    Versioning follows the Flask app's `__version__` constant in
    `app/app.py` — bump them in lockstep per
    `app/guides/RELEASE.md`.
  contact:
    name: Couchbase Query Analyzer
    url: https://github.com/Fujio-Turner/cb_completed_request
  license:
    name: See LICENSE file in repo root
    url: https://github.com/Fujio-Turner/cb_completed_request/blob/main/LICENSE
servers:
  - url: http://localhost:8888
    description: Local development (default Flask port)
  - url: http://localhost:5555
    description: Playwright E2E test port
...
paths:
...
  /api/couchbase/save-analyzer:
    post:
      operationId: saveAnalyzer
      tags: [Analyzer]
      summary: Save an analyzer report to CBL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId, analyzerData]
              properties:
                requestId:
                  type: string
                  description: Unique ID for the report (typically a UUID).
                name:
                  type: string
                  default: Untitled
                analyzerData:
                  type: object
                  description: Full analyzer report payload.
                  additionalProperties: true
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/couchbase/load-analyzer/{request_id}:
    post:
      operationId: loadAnalyzer
      tags: [Analyzer]
      summary: Load an analyzer report by request ID
      parameters:
        - name: request_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Loaded report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzerReport"
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/couchbase/delete-analyzer:
    post:
      operationId: deleteAnalyzer
      tags: [Analyzer]
      summary: Delete a saved analyzer report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId]
              properties:
                requestId: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  # ── Preferences ───────────────────────────────────────────────────────────
  /api/couchbase/save-preferences:
    post:
      operationId: savePreferences
      tags: [Preferences]
      summary: Save per-user UI preferences to CBL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, preferences]
              properties:
                userId:
                  type: string
                  description: User identifier (typically "default" for single-user installs).
                preferences:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/couchbase/load-preferences/{user_id}:
    post:
      operationId: loadPreferences
      tags: [Preferences]
      summary: Load per-user UI preferences
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Preferences object (may be empty for new users)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  # ── AI ────────────────────────────────────────────────────────────────────
...
  /api/ai/status/{document_id}:
    post:
      operationId: aiStatus
      tags: [AI]
      summary: Poll the status of a running AI analysis
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current status of the AI run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIHistoryEntry"
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/cancel:
    post:
      operationId: aiCancel
      tags: [AI]
      summary: Cancel a running AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id]
              properties:
                document_id: { type: string }
      responses:
        "200":
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/ai/history:
    post:
      operationId: aiHistory
      tags: [AI]
      summary: List AI analysis history (paginated)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                clusterName: { type: string }
                limit:  { type: integer, default: 50 }
                offset: { type: integer, default: 0  }
      responses:
        "200":
          description: Paginated history
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AIHistoryEntry" }
                  total:  { type: integer }
                  limit:  { type: integer }
                  offset: { type: integer }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/clusters:
    post:
      operationId: aiClusters
      tags: [AI]
      summary: List distinct cluster names that appear in AI history
      responses:
        "200":
          description: Cluster names
          content:
            application/json:
              schema:
                type: object
                properties:
                  clusters:
                    type: array
                    items: { type: string }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/stats:
    get:
      operationId: aiStats
      tags: [AI]
      summary: Aggregate AI usage stats from CBL ai_history
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/payload-reference:
    get:
      operationId: getPayloadReference
      tags: [AI]
      summary: Read the current AI payload-reference document from CBL
      responses:
        "200":
          description: Payload reference
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/payload-reference/load:
    post:
      operationId: loadPayloadReference
      tags: [AI]
      summary: Load the payload-reference document explicitly into memory
      responses:
        "200":
          description: Loaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/ai/payload-reference/seed:
    post:
      operationId: seedPayloadReference
      tags: [AI]
      summary: Seed the payload-reference doc from payload_reference.json.template
      responses:
        "200":
          description: Seeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/ai/payload-reference/save:
    post:
      operationId: savePayloadReference
      tags: [AI]
      summary: Persist the payload-reference document to CBL
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/ai/models:
    get:
      operationId: getAiModels
      tags: [AI]
      summary: Read the AI provider/model registry from CBL
      responses:
        "200":
          description: Models list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/ai/models/seed:
    post:
      operationId: seedAiModels
      tags: [AI]
      summary: Seed the models list from ai_models_list.json.template
      responses:
        "200":
          description: Seeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/ai/models/save:
    post:
      operationId: saveAiModels
      tags: [AI]
      summary: Persist the models list to CBL
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }
...
  /api/storage/export:
    get:
      operationId: storageExport
      tags: [Storage]
      summary: Stream a tar.gz backup of the CBL database
      responses:
        "200":
          description: Tarball download
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
      additionalProperties: true
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Document not found" }
        code:    { type: integer, example: 404 }

    # Domain
...
  responses:
    BadRequest:
      description: Invalid request (missing field, bad JSON, bad path parameter)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            success: false
            error: "name is required"
            code: 400
...